ci: fixes
